Added a title header utility component
